R219-016 Simplify handling of constants
The rewrite of constants in the AST introduced for fixed-points is not
critical for provability, contrary to what was felt then. Better to do
without this problematic rewrite which causes issues in flow analysis
and marking.

* gnat2why/spark/spark_rewrite.adb
(Rewrite_Identifier): Remove procedure and call to it.

* gnat2why/why/why-gen-scalars.adb
(Generate_Range_Predicate): Use Expr_Value_R instead of Realval as bound
 or type might not be a literal anymore.

* /M809-005__float_basics/proof/sessions/bar/why3session.xml
* /M809-005__float_basics/proof/sessions/foo/why3session.xml
* /M809-005__float_basics/proof/sessions/homothetical/why3session.xml
Update sessions.

* testsuite/gnatprove/tests/M823-017__fixed_points/test.out
* testsuite/gnatprove/tests/M823-017__fixed_points/test.py
Update script for no loss in provability.

* testsuite/gnatprove/tests/O429-042__stdlib_tasking/test.out
One check unproved because constant is in SPARK_Mode Off code.

* testsuite/gnatprove/tests/OA21-028__failsafe/test.out
* testsuite/gnatprove/tests/OA21-028__failsafe/test.py
Update script for gain in provability.

* testsuite/gnatprove/tests/OA20-014__big_shift/test.out
No loss in provability.

* testsuite/gnatprove/tests/trajectory_computation/test.out
Accept minor loss of proved checks on this complex test.

Change-Id: I52b92cd17b3c8dedfee8868b2f2aa0cc40fd7a70