R322-042 fix handling of long array aggregates
For R222-012 a guard was introduced for axioms of array aggregates.
It is formed of the conjunction of the dynamic properties of each
element. The conjunctions were chained, resulting in a deep tree.
In the new version, we still produce the same guard, but now we use
the existing list construct for that.

* gnat2why-expr.adb
(Transform_Aggregate): Instead of chaining using New_And_Expr directly,
  store conjuncts in array and then call New_And_Expr

Change-Id: I9437b3ef0df230c9c5281ccb13e74963062cdd7c