R116-018, R322-014 Reject intrinsic operator declared SPARK_Mode=>Off
When a unary or binary operator is declared with SPARK_Mode=>Off, and is
defined as an intrinsic, the use of this operator in SPARK code was not
rejected and could lead to a crash. Now fixed.

* gnat2why/spark/spark_definition.adb
(Mark_Unary_Op, Mark_Binary_Op): Detect the case to reject.

* testsuite/gnatprove/tests/R116-018__binary_op/test.out
* testsuite/gnatprove/tests/R322-014__intrinsic/test.opt
* testsuite/gnatprove/tests/R322-014__intrinsic/test.out
Tests pass now.

Change-Id: Ia718b40113d3d1ad883848a973ed17c980122a24