minor add sanity preconditions to Graph routines
Currently when graph routines are called with illegal parameters we are
getting hard to understand crashes, e.g. when DFS is called with Null_Vertex
we get a "constraint check" failure somewhere deep in that routine's body.

This patch adds obvious preconditions that give easy to understand crashes.

(no-tn-check)

* gnat2why/flow/graphs.ads
Add preconditions for Null_Vertex as parameters.

Change-Id: If6aea64ea3eab13575336745769e0c0f8b09d7c0