R323-032 concatenation on dynamically constrained arrays types
When concatenating two arrays whose root type is dynamically
constrained, we incorrectly compute the low bound of the
concatenation. Indeed, we assume that, like for unconstrained
types, it is the low bound of the left operand while we should use
the low bound of the root type instead.
As far as I can tell, this will always lead to spurious range check
might fail messages.

* gnat2why/why/gnat2why-expr.adb
(Transform_Concatenation): Use the low bound of the root type as
 the first index of the result of the concatenation.

* tests/R323-032__dynamic_concat
New test.

Change-Id: I849abc37ecd89a0bbde5053b73ed7761d879d733