R327-026 ignore unknown pragmas outside of the unit's entity
Unknown pragmas appearing outside of the unit's entity are currently
ignored, but this behaviour is buried into the code. This commit makes it
more explicit.

Also, the implicit intention in marking is that Violation_Detected flag
should be False before and after marking of a compilation unit. This will
be enforced by an assertion in a follow-up patch.

* gnat2why/spark/spark_definition.adb
(Mark_Compilation_Unit): restore Violation_Detected after marking
 Context_Items.

Change-Id: I13ad3aeddf8ffb1b152842d8fb8013777221b4cc