R314-016 harmonize calls to Is_Volatile_Function
The frontend query Is_Volatile_Function does not exactly what we
want in SPARK as it also returns True on protected functions which
are handled specifically. To work around this in proof, we were
checking whether the function was inside a protected type. In flow,
we were checking whether the function has the Volatile_Function
pragma. To harmonize this handling, we now check for the pragma in
proof too.

* gnat2why/why/gnat2why-expr.adb
(Transform_Function_Call): Check for the pragma to see whether the
 function is volatile.

* gnat2why/why/gnat2why-subprograms.adb
(Generate_Axiom_For_Post)
(Generate_Subprogram_Program_Fun)
(Translate_Expression_Function_Body)
(Generate_Subprogram_Spec): Check for the pragma to see whether the
 function is volatile.

* gnat2why/why/gnat2why-util.adb
(Use_Guard_For_Function):  Check for the pragma to see whether the
 function is volatile.

Change-Id: I3de622418d0ff080b93adf327dff98d7acf3297a