R329-025 Reject imprecise type conversion between fixed-point and integer
When the small of the fixed-point type is not an integer or the reciprocal
of an integer, the result is not guaranteed to be in the "perfect result
set" according to Ada RM G.2.3(24), and thus should be rejected by
GNATprove.

* gnat2why/spark/spark_definition.adb
(Mark): Reject imprecise conversion between fixed-point and integer.

* testsuite/gnatprove/tests/O813-018__fixed_points/fraction_fixed.adb
* testsuite/gnatprove/tests/O813-018__fixed_points/test.out
Update test.

* testsuite/gnatprove/tests/R329-025__fixed_point/fpint.adb
* testsuite/gnatprove/tests/R329-025__fixed_point/test.out
* testsuite/gnatprove/tests/R329-025__fixed_point/test.py
New test.

Change-Id: I1c07f38f3fc536320945b6dcc0b52a32ddfb8cc0