R329-054 Fix assertion failure due to checking on fixed-points
The newly added code for checking restrictions on type conversions on
fixed-point types did call Retysp on a type before the type was marked,
which led to an assertion failure.

* gnat2why/spark/spark_definition.adb
(Mark): Mark expression before checking restrictions. Also fix a typo that
 used From_Type and To_Type in the opposite meaning. This has no influence
 on analysis results as this part of the code is symmetric in these types.

* testsuite/gnatprove/tests/Q126-021__fixed_point_div/test.out
Update output. The initial violation causes now a type not to be in SPARK
and thus the second violation on the same line is not reported.

Change-Id: Ia588be3815f7a9c096ee1aebcd94b47a285ef2eb