R329-034 Cleanup of useless theories
The theory FixedPoint_Float_Conv is useless, and should not be used in
this form in any case, as there cannot be a single conversion from
fixed-point to floating-point and conversely, only for a given value of
small.

* gnat2why/why/why-atree-modules.adb
(Initialize): Remove symbols from FixedPoint_Float_Conv for
 conversion between float and fixed.

* share/spark/theories/ada__model.mlw
(FixedPoint_Float_Conv): Remove useless theory.

Change-Id: I02ecabe70d1e87dd3d043e1b54ff7303be267ccc