MA29-016,N130-012 reenable test and remove dead pragma Warnings
This test was disabled while working on the new GG; now it is safe to
reenable it.

The pragma Warnings in levels.adb are removed, because the emitted messages
appear to be correct (most like this improvement should be attributed to
generating globals of expression functions declared in package specs as if
their bodies would appear in package bodies).

Warning about analysing unreferenced Wibble in par.adb is rightly gone,
because this subprogram might be referenced from a child package (however,
we can't tell whether there are such packages).

* testsuite/gnatprove/tests/MA29-016__flow_refinement/
Reenable; remove pragma Warnings; accept current output.

Change-Id: I229732385c2691215d82f791167db38be2889d98