R329-026 reject subtype or derived types of nonspark types
Subtypes and types derived from a type which is not in SPARK
should not be in SPARK. We reject subtype and derived type
declarations mentioning a type which is not in SPARK. Note that
this is not enough to prevent uses of such types when pulled from
withed entities.

* gnat2why/spark/spark_definition.adb
(Mark): Mark subtype indication instead of parent type entity.
(Mark_Subtype_Indication): Subtype indication may mention a type
 which is not in SPARK if its Retysp is in SPARK.

Change-Id: I0bdb45432224add239b07f81995030a8fd701eb2