O820-005 Add root cause for indirect SPARK violation messages
When a declaration/expression/statement is not in SPARK due to some
other entity not being in SPARK, the current violation message does not
explain the root cause of the violation. This is particularly annoying
when the root cause of the violation is not in code marked SPARK_Mode(On),
as in that case, no violation message at all is issued on the root cause.

This patch adds a notion of root cause that is propagated from entity to
entity by two methods:
- either when an entity A is out of SPARK due to some entity B, in which
  case it inherits the root cause for B;
- or when an entity A is out of SPARK due to another violation that
  occurred while analyzing A, in which case it inherits the last root
  cause registered in this analysis.

The effect on the testsuite is that a number of violation messages now
include a root cause. Some redundant violations also disappear due to
location of multiple violations on the same entity, or other mechanisms
in the frontend that avoid printing cascading errors.

Global variables and subprograms related to violation detection are
grouped in a local package SPARK_Violations. Currently no encapsulation is
performed, although this should be added in the future.

* common/string_utils.adb
(Standard_Ada_Case): New function to capitalize a string according to the
 standard Ada case.

* gnat2why/spark/spark_definition.adb
(Violation_Root_Cause): New hash map to store root causes, which are simply
 messages initially issued for a violation. Any node can be mapped to a
 violation, in particular entities.
(Last_Violation_Root_Cause_Node): Last node on which a violation was
 reported. Useful during analysis of an entity.
(Get_Violation_Root_Cause): Return the root cause for a node.
(Add_Violation_Root_Cause): Add the root cause for a node.
(Mark_Object_Entity): Issue violation on entity preferably.
(Mark_Type_Entity): Remove duplicate detection of violation on record
 components. Issue violation on entity preferably. Use the newly defined
 Last_Violation_Root_Cause_Node to store a violation detected during
 analysis of the entity.
(Mark_Unsupported): Remove parameter Extra_Node which was used for a single
 violation message and made root cause messages more complex.
 Add violation root cause.
(Mark_Violation): Add violation root cause. In the version propagating
 a violation from another entity, also propagate the root cause, and
 reference this root cause in the violation message.
(Mark_Attribute_Reference, Mark_Default_Expression): Call Mark_Violation
 instead of directly issuing violations, which is fragile in case of changes
 like here.

* testsuite/gnatprove/tests/K222-004__package/test.out
* testsuite/gnatprove/tests/K222-004__package_spec/test.out
* testsuite/gnatprove/tests/K222-004__subprogram/test.out
* testsuite/gnatprove/tests/K222-004__ug/test.out
* testsuite/gnatprove/tests/K302-029__ug/test.out
* testsuite/gnatprove/tests/M920-013__indirect_call/test.out
* testsuite/gnatprove/tests/M923-035__discr/test.out
* testsuite/gnatprove/tests/N605-022__discrim/test.out
* /gnatprove/tests/O416-002__dynamic_predicate_not_in_spark/test.out
* testsuite/gnatprove/tests/O429-044__synchronous_barriers/test.out
* testsuite/gnatprove/tests/O602-020__dic_not_spark/test.out
* testsuite/gnatprove/tests/O916-043__crazyflie/test.out
* testsuite/gnatprove/tests/O924-001__ext_axiom_private_violation/test.out
* testsuite/gnatprove/tests/OB20-025__marking_protected_calls/test.out
* /gnatprove/tests/OC17-029__package_body_statements_marking/test.out
* testsuite/gnatprove/tests/P111-029__marking_delayed_aspect/test.out
* testsuite/gnatprove/tests/P112-047__record_not_in_spark/test.out
* testsuite/gnatprove/tests/P208-037__interrupt_priority_in_spark/test.out
* testsuite/gnatprove/tests/P513-017__explicit_dereference/test.out
* testsuite/gnatprove/tests/P513-018__flow_non_spark_obj_decl/test.out
* testsuite/gnatprove/tests/P615-019__stone_level/test.out
* testsuite/gnatprove/tests/P706-024__private_discr_type/test.out
* testsuite/gnatprove/tests/QA02-024__dimensioned/test.out
* testsuite/gnatprove/tests/QC14-025__tasking_pointers/test.out
* testsuite/gnatprove/tests/TU__derived_types_and_classes_illegal/test.out
* testsuite/gnatprove/tests/O110-002__default_init/test.out
* testsuite/gnatprove/tests/O429-044__tasking_filter/test.out
* testsuite/gnatprove/tests/OB19-019__task_storage_size/test.out
* testsuite/gnatprove/tests/P820-002__protected_out/test.out
* testsuite/gnatprove/tests/PA03-051__flow_iternal_call_prot_init/test.out
* testsuite/gnatprove/tests/PA20-004__flow_internal_protected_init/test.out
* testsuite/gnatprove/tests/QB07-050__single_po/test.out
* testsuite/gnatprove/tests/QC05-006__machine_rounding/test.out
Output update, with root causes added, and some duplicate violation messages
removed.

* testsuite/gnatprove/tests/O210-001__attribute_deref/test.out
* testsuite/gnatprove/tests/O708-001__address/test.out
* testsuite/gnatprove/tests/O722-011__standard_attributes/test.out
* testsuite/gnatprove/tests/OB19-012__task_identity_attribute/test.out
* testsuite/gnatprove/tests/P229-020__attribute_address/test.out
Update output on errors related to attributes.

* /gnatprove/tests/O820-005__indirect_violation/max_array_dimensions.ads
* testsuite/gnatprove/tests/O820-005__indirect_violation/other.ads
* testsuite/gnatprove/tests/O820-005__indirect_violation/test.gpr
* testsuite/gnatprove/tests/O820-005__indirect_violation/test.out
* testsuite/gnatprove/tests/O820-005__indirect_violation/test.py
* testsuite/gnatprove/tests/O820-005__max_array_dimensions/test.out
New test showing the benefits of the new scheme.

* testsuite/gnatprove/tests/R323-066__violation/test.out
Better violation messages on this recent test.

Change-Id: I27107965d0e73cd639c84ef7fbd77514cfc3c684