R403-041 bring ironsides test back to large testsuite
Even when using memcached, the ironsides test still takes long to analyse;
on my setup it increases the time required to run the testsuite from 3 to 4
minutes.

Most likely this should be blamed on large .mlw files (5 of them is between
4 and 5 MB) and should be solved by existing TN, i.e. top-down
marking (P805-023) and improvements to the Why AST processing (Q306-001,
Q327-030).

* testsuite/gnatprove/tests/ironsides/test.opt
Move back to the large testsuite.

Change-Id: Iedc0a0c49bc416d43c12e848640c818e7a28afc6