QA12-018 Improve Check_Elaborate_Body
When checking for a missing pragma Elaborate_Body we were only checking
for variables directly defined in the body of a package. We are now also
checking global outputs of subprograms calls that happen in the body.

* gnat2why/flow/flow-analysis.adb
(Check_Elaborate_Body): inspect global outputs of subprogram calls as well

* testsuite/gnatprove/tests/N108-029__flow_generate_initializes_1
* testsuite/gnatprove/tests/PC19-013__flow_ghost_output
* testsuite/gnatprove/tests/TU__initial_condition_illegal
* testsuite/gnatprove/tests/TU__tasks_legal
Update test outputs

* testsuite/gnatprove/tests/QA12-018__flow_elaborate_body
Expand test and update test output. We are still missing a check for Y.
This will be fixed under another ticket.

Change-Id: I99e6307c4fc21e6260435dda7ae7138f8dd914ba