R411-005 Remove failing assertions
This removes "assert (false)" which should never be triggered. They are
actually triggered when an external axiomatization that omits model_trace
labels is added (in counterexample pass).
No test output changed as this is just the removal of assert.

* gnat2why/why/gnat2why-counter_examples.adb
(Get_Entity_Id): Remove assertion.
(Refine_Aux): Prepend condition on presence of Field_Entity.

Change-Id: Ib84226ff337ff686bcd5f81326a9e4d17e86da2f