R413-022 Fix issues with access to volatile arrays in proof
Fix access to a volatile variable, when either indexing into a volatile
array, or accessing attributes First, Last and Length. This requires that
the sequence of Why nodes representing a volatile variable in programs
(havoc var; var) is handled correctly in code that expects to retrieve
the Ada node from the sequence Why node.

* gnat2why/why/gnat2why-expr.adb
(Transform_Identifier): Pass on the Ada node to the sequence that
 represents the entity.

* gnat2why/why/why-gen-arrays.adb
(Get_Entity_Of_Variable): Deal with Why node as a sequence to represent
 an array variable.

* gnat2why/why/why-gen-progs.adb
* gnat2why/why/why-gen-progs.ads
(Sequence): Add a version that accepts an Ada node.

* testsuite/gnatprove/tests/R411-018__volatile_length/p.adb
* testsuite/gnatprove/tests/R411-018__volatile_length/p.ads
* testsuite/gnatprove/tests/R411-018__volatile_length/test.out
* testsuite/gnatprove/tests/R411-018__volatile_length/test.py
New test.

Change-Id: I38ce7e24e594a63e9ed9d94464db637c26e009d6