R412-049 correctly handle references to discriminals
Sometimes the frontend will use in parameters (discriminals) to
refer to discriminants of an object inside the object itself. These
renamings should not be translated in SPARK, instead the
discriminant should be used directly. This was already done
properly for concurrent type, but it is also necessary for regular
records. Indeed, discriminals will be used for occurrences of their
discriminant in default expressions of record components for
example.

* gnat2why/why/gnat2why-driver.adb
(Translate_Entity): Discriminals of regular records should not be
 translated either.

* gnat2why/why/gnat2why-expr.adb
(Transform_Identifier): Discriminal can also occur for regular
 record types. Remove incorrect assertion.

Change-Id: I69d8f5a5ce366aa8e85743b55016e097981bdd5c