accept current output on ironsides test
The ironsides test was recently touched: enabling, adding some SPARK_Mode =>
Off, removing and restoring the 'large' qualifier. Apparently, we lost track
of the current output, which was missed due to rlimit timeouts on nightly
builds. This patch merely accepts the current output as it is to give us a
stable reference.

(no-tn-check)

* testsuite/gnatprove/tests/ironsides/test.out
Update.

Change-Id: Ic4cd008377418dbfcb4eb8c55d3d8ae0eaee8db5