R416-004 improve frame condition on array slices
We improve here the case of an iteration over a slice of an array.
We used to generate a frame condition stating that indexes which
are both outside of the slice and bigger than the loop index are
preserved. We rather want to express that indexes which are either
outside of the slice or bigger than the loop index are preserved.
There are two parts in this improvement. First, we now ignore
slices which are prefixes of other slices or array accesses. This
avoids generating the conjunction above. Then, we assume that
indexes outside the range of the loop are also preserved.

* gnat2why/why/gnat2why-expr-loops-inv.adb
(Equality_Of_Preserved_Components): New parameters Low_Id and
 High_Id for the low and high bounds of the loop. If the accessed
 indice is the loop index, then indices outside the loop range are
 preserved.
(Generate_Frame_Condition): New parameters Low_Id and High_Id for
 the low and high bounds of the loop if any.
(Update_Status): New parameter Ignore_Slice to ignore slices in the
 prefix of a slice or an indexed component.

* gnat2why/why/gnat2why-expr-loops.adb
(Transform_Loop_Statement): Give low an high bound of the loop to
 Generate_Frame_Condition if any.

* tests/R416-004__frame_condition
New test for preservation outside of the loop range.

* tests/Q719-013__loop_unroll_dyn/test.out
* tests/QB10-026__counterexample/test.out
Update counterexamples.

*  tests/riposte__array_applications
Update sessions.

Change-Id: Idc8ecf7f113d4b8c114b0da7d6848f0bc19a7a08