Q113-013 Improve messages on constituents
Messages about constituents could be ambiguous, in particular, when
constituents declared in different packages have the same names.
We now append the immediate scope of the constituent when it differs
from the immediate scope of the abstract state.

* gnat2why/flow/flow_error_messages.adb
(Substitute): add extra comment

* gnat2why/flow/flow_types.adb
(Flow_Id_To_String): for constituents whose immediate scope differs than
the immediate scope of their enclosing state, print the immediate scope as
a prefix.

* testsuite/gnatprove/tests/Q113-013__flow_constituents
Add test.

* testsuite/gnatprove/tests/M314-014__flow_packages
Restore original test and update output. It is now easier to understand to
which constituents the message is referring to as they are prefixed with
their immediate scope.

* testsuite/gnatprove/tests/M729-011__flow_body_stubs
* testsuite/gnatprove/tests/MA29-016__flow_refinement
* testsuite/gnatprove/tests/MB26-023__flow_nested_packages_effects
* testsuite/gnatprove/tests/OC08-044__refinement
* testsuite/gnatprove/tests/Q728-005__type_invariants_private_child
* testsuite/gnatprove/tests/Q823-011__flow_hidden_var_in_priv_spec
* testsuite/gnatprove/tests/QB10-014__flow_priv_child
* testsuite/gnatprove/tests/RM_MS__asm_abstract_state_refined_in_embedded_
and_private_child
* testsuite/gnatprove/tests/RM_MS__asm_abstract_state_refined_in_embedded_
package
* testsuite/gnatprove/tests/prime_numbers

Change-Id: If1dd4a6a3ca502460413a6abb8272c89094ae932