PC02-011 Marking of private declarations of a package
Right now we mark the private declarations of a package when the private
part has a SPARK_Mode => Off. Ideally we shouldn't. This is a workaround
this. In case the SPARK_Mode of the private part is Off then we restore
the value for Violation_Detected that we had before the marking of the
private declarations.

* gnat2why/spark/spark_definition.adb
(Mark_Package_Declaration): ignore Violation_Detected status if after the
marking of private declarations that are behind a SPARK_Mode => Off.

* testsuite/gnatprove/tests/N108-029__flow_generate_initializes_2
Test output updated.

* testsuite/gnatprove/tests/PA30-001__part_of_spark_mode_off
Enable test as it does not crash anymore. Add test output.

Change-Id: I479a1ee9328db84479187b2119fce70dcee5d09e