R423-012 New warning on dead code after loop
This new warning signals when the code after a loop is unreachable.
This might be either because the loop never exits normally, or because
the loop itself is unreachable. Special care is taken not to issue
the warning on statements meant to signal abnormal execution, like
a raise statement.

This warning is triggered by switch --proof-warnings and can be
suppressed like other warnings. It is relying on a call to provers, like
the previously implemented warning on inconsistent preconditions.

* common/vc_kinds.ads
(VC_Kind): New VC kind for dead code.

* gnat2why/spark/spark_util.adb
* gnat2why/spark/spark_util.ads
(Is_Error_Signaling_Statement): New function to detect error-signaling
 statements.

* gnat2why/why/gnat2why-error_messages.adb
(Proved_Message): Support new VC kind.

* gnat2why/why/gnat2why-expr-loops.adb
(Wrap_Loop): Generate VC for warning on statement after the loop when
 appropriate.
(Transform_Loop_Statement): Pass new argument Next_Stmt to Wrap_Loop.

* gnat2why/why/why-gen-preds.ads
Fix layout.

* testsuite/gnatprove/tests/K330-014__loop_entry/test.out
* testsuite/gnatprove/tests/K330-014__loop_entry/test.py
* testsuite/gnatprove/tests/M213-022__flow_loops/test.out
* testsuite/gnatprove/tests/M213-022__flow_loops/test.py
* testsuite/gnatprove/tests/P715-007__termination/test.out
* testsuite/gnatprove/tests/P715-007__termination/test.py
* testsuite/gnatprove/tests/P715-019__flow_subprogram_termination/test.out
* testsuite/gnatprove/tests/P715-019__flow_subprogram_termination/test.py
* testsuite/gnatprove/tests/PC08-013__flow_funcs_inf_exec/test.out
* testsuite/gnatprove/tests/PC08-013__flow_funcs_inf_exec/test.py
* testsuite/gnatprove/tests/Q512-013__flow_type_aspects/test.out
* testsuite/gnatprove/tests/Q512-013__flow_type_aspects/test.py
Issue new warning on various tests.

* testsuite/gnatprove/tests/P508-001__bits_manipulation/test.out
* testsuite/gnatprove/tests/P508-001__bits_manipulation/test.py
Particularly interesting case where the warning is quite subtle, as the loop
is indeed infinite due to the combination of V being only assigned non-zero
values in the loop, and the loop test checking for V being null to exit.

* testsuite/gnatprove/tests/R423-012__dead_code/dead_code.adb
* testsuite/gnatprove/tests/R423-012__dead_code/dead_code.ads
* testsuite/gnatprove/tests/R423-012__dead_code/test.out
* testsuite/gnatprove/tests/R423-012__dead_code/test.py
New test.

Change-Id: I71cb92bb9513d7b67853a948cd1763dcd8ebb654