R410-023 Properly check components and count field of record ce
For a type T and its subtype S, some components of T were previously
returned as components of a counterexample of S. This commit solves the
issue.
It also now properly counts fields so that (others => ?) is printed only
when necessary.

* gnat2why/why/gnat2why-counter_examples.adb
(Refine_Aux): CNT_Record case: Add a check for fields so that only
discriminant and true component of the type are printed.

* gnat2why/why/gnat2why-ce_util.ad*
Added utils specific to counterexamples pass.
(Count_Why_Visible_Regular_Fields): Add function.
(Is_Visible_In_Type): Add function.

* testsuite/gnatprove/tests/L531-021__discr/test.out
The counterexample is what we wanted to obtain.

Change-Id: I71db4f65162a7601f4e35597f29520323831d6e6