R426-023 Add test
This test checks that gnatwhy3 does not fail when there are two checks
with the exact same reason and location.

* testsuite/gnatprove/tests/R426-023__duplicate_vc/main_range.mlw
Example file: the vc with "VC_RANGE_CHECK" reason is duplicated.

* testsuite/gnatprove/tests/R426-023__duplicate_vc/test.out
Results for the different VC as given by gnatwhy3.

* testsuite/gnatprove/tests/R426-023__duplicate_vc/test.py
This execute gnatwhy3 with cvc4 prover and with or without debug mode.
It then parses the JSON output from gnatwhy3 and give only the id with
their corresponding result (true or false). The association ID, Result
should never change.

* testsuite/gnatprove/tests/R426-023__duplicate_vc/why3.conf
Minimal why3.conf.

Change-Id: If8a5efba3c457f2969dc5b5b795cc77e734197ee