R206-018 Update test
This changes the behavior of this test which would formerly try to run
gnatprove on a session with existing manual proof (but obsolete): This
should not work but a warning should be emitted as part of R219-007.
For this test to work, we first replay the obsolete session. This is what
we want the users to do and what will be documented.

* testsuite/gnatprove/tests/R206-018__itp_mix/proofs/../why3session.xml
Update session. A lot of theory removed probably thanks to topdown marking

* testsuite/gnatprove/tests/R206-018__itp_mix/test.py
Add replay call

* testsuite/gnatprove/tests/R206-018__itp_mix/test.o*
Update test.

Change-Id: I3cd47123109ef10b14e911bcad69c6f253a6b22f