R427-016 checks on store for actual parameters
Clean up checks on parameters on return of subprogram calls. Use
Lvalue parameter to chose the type to check on parameters of scalar
types. Do not special case actuals which are conversions. Perform
checks on actuals which are arrays in split form even if the
reference can be used as is. Perform predicate check even on
actuals which are conversions of non-scalar types and even when no
references is necessary.

* gnat2why/why/gnat2why-expr.adb
(Compute_Call_Args): A check can be needed on an actual of an array
 type in split form. Insert a conversion and discard the result to
 reuse the reference in this case.
(Insert_Ref_Context): We may still need to insert predicate checks
 even if no ref is needed, so make Nb_Of_Refs a natural. Remove
 special handling of conversions on copy back of regular
 parameters.
(New_Assignment): Introduce checks for conversions in left hand
 side.
(New_Predicate_Check): Move declaration to ads so that it can be
 used in why-gen-expr.
(Transform_Statement_Or_Declaration): Call Insert_Ref_Context even
 if no references are needed to get the checks on store for
 predicates.

* gnat2why/why/why-gen-expr.ads
(Insert_Array_Conversion): Reconstruct the expression before
 checking the predicate to avoid crashing on arrays in split form.
(Insert_Scalar_Conversion): Use Lvalue parameter to decide if
 parameters should be checked against the formal or the actual
 type. Do not include predicate check when no Ada_Node is given
 (when called from Insert_Simple_Conversion).

* tests/M318-008__flow_generics/test.out
Spurious range checks which where not introduced before for bad
reasons (check on array in split form).

* tests/O630-015__modular_conversion/test.out
* tests/P620-034__float_out_subtype/test.out
Locate check on store on argument of conversion.

* tests/R427-016__check_on_param_call
New test for checks on actuals.

* tests/pair_insertion_sort/test.out
Spurious predicate checks which where not introduced before for bad
reasons (subprogram has no reference).

Change-Id: I91b40202170be62901e51052d2146a09db933e45