R503-009 correct failing assertion in developer build
In the previous patch, the assumption was made that why types of
EW_Abstract or EW_Split kind always have Ada nodes. This assumption
is broken in one particular place of gnat2why. Dummy types are
created for private parts of types whose fullview is not in SPARK.
These types have kind EW_Abstract and no Ada nodes. We still call
Insert_Single_Conversion on them though conversion on these dummy
types is not defined. It is OK though as no conversion is actually
required, to and from types match.

* gnat2why/why/why-inter.adb
(Eq_Base): Support being called on dummy types of kind EW_Abstract
 with no Ada nodes, but only if the type names match.

Change-Id: I1edd38f2ba2315223cbcb78a48785fb2d4a5ee84