R514-013 check type invariant on first subtype
The frontend sometimes introduces an Itype as a base type of a type
with an invariant. In this case, the invariant is also specified on
the base itype. We used to check that the invariant is free of
runtime errors on this base itype, as it is the first type in the
hierarchy with the invariant. However, it is not appropriate as the
base itype may miss some constraints which would make the invariant
correct.
We now check the invariant on the first subtype instead.

* gnat2why/why/gnat2why-expr.adb
(Check_Type_With_Invariants): Check the invariant on the first
 subtype.

* tests/R514-013__invariant
New test.

Change-Id: Ib9e42a35e3fcc9a6ee79c0d8b7d408fc53878346