R515-031 Fix marking of bounds of scalar types
Bounds of scalar types (both user-defined and implicit ones) should be
in SPARK for the type to be in SPARK. This is now taken into account
in marking. Otherwise some crashes are possible on code not marked in or
out of SPARK.

* gnat2why/spark/spark_definition.adb
(Mark_Type_Entity): Check that the range of types is in SPARK. Only do
 so when the scalar type has not been already marked as unsupported, in
 order to keep the most precise diagnostic message if present.

* testsuite/gnatprove/tests/P805-023__array_index/test.out
* testsuite/gnatprove/tests/R326-021__component/test.out
Update output.

Change-Id: Ib151502200195f1cc6d3ff3099897f2fbcc15d91