P202-036 all N_Integer_Literals are now decorated with an Etype
It looked like an oversight in the implementation of the Max_Queue_Length
pragma that pragma expression was translated into an integer literal without
an Etype. This is now fixed in the frontend and consequently all numeric
literals that we see in the backend have a non-empty Etype.

* gnat2why/spark/spark_definition.adb
(Mark): don't expect empty Etype of numeric literals.

Change-Id: Ia0649d0d7221f5f7616122cb2ab65f69450882f5