QB29-037 only check flow traces on flow-related tests
When running the testsuite with --cache we only get traces emitted from flow
and don't get those emitted from proof. This patch adapts the test drivers
so that in tests related to traces emitted from flow we ignore all the other
traces.

* testsuite/gnatprove/lib/python/test_support.py
(check_trace_files): remove "opt" option, which it was outdated and no
 longer used; add flow_only option.

* testsuite/gnatprove/tests/M213-042__flow_error_messages/
Adapt test driver and test output.

* testsuite/gnatprove/tests/RM_Examples/test.py
Narrow the selection of traces included in test output (so that if proof
emits some traces in the future they will be ignored).

Change-Id: I29500845f25f2ccb1ed5729da70389e52cabf029