fix numbering of LRM trace unit labels in tasking chapter
In all other chapters the trace unit labels coincide with the rule numbers.

Also, it is confusing when we reference a given rule in emails sent to the
customers with an URL which uses a different number (as it happened on
R530-005).

* docs/lrm/source/tasks-and-synchronization.rst
Make trace units labels match the rule numbers.

Change-Id: I23adbee3a78bd883368f3eafc052e91d7fcc333d