R601-033 crash on update of nested discriminant dependent component
The frontend may introduce an Itype for the type of a discriminant
dependent component of a record in an assignment. This type causes
a mismatch in the generated Why3 code. Change gnat2why to compute
the correct type for the prefix of an assignment instead of relying
on the frontend.

* gnat2why/why/gnat2why-expr.adb
(Expected_Type_Of_Prefix): Recursive call to get the correct type
 from the updated variable. Return an Ada node instead.
(One_Level_Update): Call Expected_Type_Of_Prefix to get the correct
 type.

Change-Id: I2de900ea0ae29c7d55b3d4d433e3b31d36e3997e