R531-025 Add lemmas for division of fixed-point value by integer
With GNAT, division of a fixed-point value by an integer follows the rules
of integer division. Add specific lemmas in the SPARK lemma library that
spell out some of these properties.

* include/spark-fixed_point_arithmetic_lemmas.adb
* include/spark-fixed_point_arithmetic_lemmas.ads
Add 3 lemmas about monotonicity of division wrt numerator, antimonotonicity
of division wrt numerator when all arguments are non-negative, and division
being the inverse of multiplication. This is only on division of a
fixed-point value by an integer value (here always positive).

* testsuite/gnatprove/tests/R531-025__fixed_arithmetic_lemmas/test.gpr
* testsuite/gnatprove/tests/R531-025__fixed_arithmetic_lemmas/test.out
* testsuite/gnatprove/tests/R531-025__fixed_arithmetic_lemmas/test.py
* /gnatprove/tests/R531-025__fixed_arithmetic_lemmas/test_fixed_points.adb
* /gnatprove/tests/R531-025__fixed_arithmetic_lemmas/test_fixed_points.ads
Add test for new lemmas.

Change-Id: I51dfbfcdc398dd892aa4b9156ee4e9a84461ee27