R123-048 Fix analysis of code in SPARK_Mode Auto
When SPARK_Mode Auto is used on some tasking unit, and the correct SPARK
configuration is not set, a protected subprogram could end up being marked
for translation while the enclosing protected type was marked as not in
SPARK. This could lead to a crash in proof. Now fixed.

One subtlety of the current implementation is that the respect of tasking
profile in SPARK is based on the values of pragmas picked for the main
unit being translated. This is more liberal than SPARK RM 9(2) rule.

* gnat2why/spark/spark_definition.adb
(Mark_Subprogram_Entity): Mark violation on protected operation if not in
 the correct configuration for SPARK tasking.

* testsuite/gnatprove/tests/R123-048__tasks_simplified/test.opt
* testsuite/gnatprove/tests/R123-048__tasks_simplified/test.out
* testsuite/gnatprove/tests/R123-048__protected_auto/test.opt
* testsuite/gnatprove/tests/R123-048__tasks/test.opt
* testsuite/gnatprove/tests/R123-048__tasks/test.out
Tests pass now.

Change-Id: Ia42c7c3347c4d9596f745cd73791fbf5b4cb24e6