P805-023 improve top-down marking of predicates
The "core" of this commit is to also mark bodies of predicate functions
from Mark_Subprogram_Entity (the code does it for all predicate
functions, though we are really only interested in those that are in
withed units).

Just as for expression functions, this requires skipping some calls to
Mark_Subprogram_Body (and in this case also to
Mark_Subprogram_Declaration) when traversing the current unit via
Mark().

Finally, with these changes we would miss one case of a private type
with predicates whose full view is not in SPARK (this happens on
P112-038__predicate_private_type_not_in_spark). We check for this case
explicitly, which also improves the error message in two cases.

* spark_definition.adb
(Mark): protect call to Mark_Subprogram_Body a bit better; skip a call
  to Mark_Subprogram_Declaration
(Mark_Subprogram_Entity): also mark bodies of predicate functions
(Mark_Subprogram_Type): explicit code to check for error condition

Change-Id: I5fc168b2ff8b85610724a98e905f32400defbb8b