QB10-026 array types which aren't statically constrained are clones
Do not introduce a new clone of Unconstr_Array for derived type and
subtypes which are not statically constrained, but instead reuse
the clone introduced for the base type.

* gnat2why/why/gnat2why-types.adb
(Translate_Type): Array types can be clones too.

* gnat2why/why/why-gen-arrays.adb
(Array_Type_Cloned_Subtype): The array type which has been cloned.
 This is always the root of the derivation tree for now.
(Array_Type_Is_Clone): Return True for types which are not
 statically constrained and which are not roots of their derivation
 tree.
(Declare_Ada_Array): Clone the appropriate subtype if the type is a
 clone.

* testsuite/gnatprove/tests/R509-011__higher_order_proof
Update session test.

* testsuite/gnatprove/tests/P630-037__keccak/test.out
* testsuite/gnatprove/tests/Q930-002__itp/test.out
A new proof.

* testsuite/gnatprove/tests/QB10-026__counterexample/test.out
A proof is lost but it used to go through by mistake (the test is
largely unproved and a similar check at the following line was
already unproved).

* testsuite/gnatprove/tests/KB25-022__exp/test.out
* testsuite/gnatprove/tests/L831-037__slice/test.out
* testsuite/gnatprove/tests/M830-004__checks/test.out
* testsuite/gnatprove/tests/N622-004__bitwise_ops/test.out
* testsuite/gnatprove/tests/NA11-009__very_long/test.out
* testsuite/gnatprove/tests/NA11-009__very_long_subunit/test.out
* testsuite/gnatprove/tests/R125-020__comparison/test.out
* testsuite/gnatprove/tests/R419-026__swap_ranges/test.out
Counterexamples only.

Change-Id: I2c5aac8988cae2f2e741a7771ceddef984a9f8ef