QA11-077 temporarily accept diff related to discriminants
This diff comes from removing a broken handling of discriminants in
Get_Globals (which is the subject of this TN); it will be fixed later.

Previously, when processing the AST of the the Global/Refined_Global
contract we special-cased discriminants in a way that had no justification
in the SPARK RM. This was controlled by Consider_Discriminants parameter.

Now, processing of the AST is refactored into Parse_Global_Contract (to
reuse it outside of Get_Globals) and the Consider_Discriminants parameter
is marked as unreferenced.

* testsuite/gnatprove/tests/O827-017__dispatching_procedures/test.out
Update output (not sure whether this is regression or improvement).

Change-Id: I117da256c1c4a2a65a62a6fea34539bcfe9da5c1